+++ Makefile
@@ -22,5 +22,6 @@ package:
 	python setup.py bdist_wheel
 
 test:
+	python test/integration/start_rabbitmq_broker.py
 	coverage run --omit src -m pytest -vv --timeout 30
 	coverage html
